requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.3
django-ckeditor==5.4.0
django-js-asset==0.1.1
idna==2.6
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
sparkpost==1.3.5
urllib3==1.22
